Export ByteScanner, remove scanner() function (unused)
ByteScanner should be obvious when it's necessary. I'll probably
implement some sort of decoder-like type so it can do buffering on its
own, and then ByteScanner can be lowered to probably just io.Reader
(can't add more requirements to it, so it's more restrictive right
now).

scanner function was unused. May be used in the server, but that's not
yet started.