Add fred/resv -- contains a basic RESP server
Server currently handles basic Resps sent to it. There's nothing
special like muxers or anything, so it's pretty dumb right now.

How Handler is defined is likely to change, since I'll probably find
some way to pass a connection ID through to the handler so that the
handler can identify itself somehow. The ID will only be unique to the
server, though, and probably just an incremented uint64. So, on the
chance that you have over 2^64 connections, it may rollover and you
might end up with some reused IDs. Don't know yet.